+++ apptentive-android-sdk/src/com/apptentive/android/sdk/storage/MessageStore.java
@@ -24,4 +24,6 @@ public interface MessageStore extends PayloadStore {
 	public String getLastReceivedMessageId();
 
 	public int getUnreadMessageCount();
+
+	public void deleteAllMessages();
 }